refactor: final code cleanup
