+++ contracts/staking/src/lib.rs
@@ -9,6 +9,7 @@ pub mod msg;
 pub mod oracle;
 pub mod query;
 pub mod state;
+pub mod types;
 
 #[cfg(test)]
 pub mod tests;
