+++ contracts/staking/src/query.rs
@@ -8,11 +8,11 @@ use crate::state::{
     unstake_requests, UnstakeRequest, BATCHES, CONFIG, IBC_WAITING_FOR_REPLY, INFLIGHT_PACKETS,
     PENDING_BATCH_ID, STATE,
 };
-use cosmwasm_std::{Deps, Env, StdResult, Timestamp, Uint128};
+use cosmwasm_std::{Deps, StdResult, Timestamp, Uint128};
 use cw_storage_plus::Bound;
 use milky_way::staking::{Batch, BatchStatus};
 
-pub fn query_config(deps: Deps, _env: Env) -> StdResult<ConfigResponse> {
+pub fn query_config(deps: Deps) -> StdResult<ConfigResponse> {
     let config = CONFIG.load(deps.storage)?;
 
     let res = ConfigResponse {
